Makes model a package (Note: run pip install -r requirements.txt)
To fix import and dependency issues, this commit makes the model directory a package called socialforce.
It must be installed using pip install -r requirements.txt (from the directory containing requirements.txt).
This allows other (independent) directories/modules to exist using the socialforce module, even when running scripts from those directories (e.g. batch_run.py).
Note that the socialforce module is installed using the flag -e, which allows for modification of the source files in this project.
Changes made to this source files are reflected in other files upon next import (or for example next run of the script)